requirements.txt
tensorflow-cpu==2.16.1
numpy
pillow
matplotlib
scikit-learn
streamlit

